Add WindowOrWorkerGlobalScope
Additionally, this fixes the following:

* IDBEnvironment removed because it's officially obsolete and "although
  it may still work in some browsers, its use is discouraged since it
  could be removed at any time".

* indexedDB now wrapped in js.UndefOr because it's not always available

* indexedDB added to WorkerGlobalScope

* caches return type changed from js.Any to js.UndefOr[CacheStorage]

* caches added to Window

* CacheStorage moved into experiment.cachestorage as it's available from
  Window too

* Add missing methods to window and web-worker scopes:
  * crossOriginIsolated
  * isSecureContext
  * origin

Ref: https://developer.mozilla.org/en-US/docs/Web/API/WindowOrWorkerGlobalScope
